statistics/pdf/requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
fpdf>=1.7.2
seaborn>=0.11.0
numpy>=1.20.0
Pillow>=8.0.0 